Clean up file organisation in eudemo.ivc.panelling
